Add #20 옹알이(1)
1.replace 함수를 통해 조카가 말할 수 있는 단어들은 -로 바꿔준다.
2.하이픈으로 대체하는 이유는 wyeoo 같은 경우 ye를 제거후 woo를 제거할 수 있기 때문에 공백을 넣으면 제대로 파악할 수 없기 떄문이다.
3.2중 반복문중 안쪽 반복문이 끝나고 해당 i번째의 단어가 -로 바뀌는 작업이 끝났을때 남은 replaceAll을 이용해 하이픈을 전부 제거후 "" 값이 남았다면 카운팅.
